requirements.txt
future==0.16.0
gensim==3.2.0
google-auth==1.2.1
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
grpcio-tools==1.8.1
h5py==2.7.0
Keras==2.1.2
librosa==0.5.1
nltk==3.2.5
numpy==1.13.3
protobuf==3.5.0.post1
scipy==1.0.0
SoundFile==0.9.0.post1
tensorflow==1.2.0
Theano==1.0.1
flask==0.12.2
python-socketio==1.8.4
eventlet==0.21.0
pandas==0.22.0
